+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Renamed NSURL `-queryDictionary` to `-queryParameters`, `-URLByAddingQueryDictionary:` to `-URLByAddingQueryParameters:`.
 - Renamed NSString `-queryDictionary` to `-URLQueryParameters`, `-stringByAddingQueryDictionary:` to `-stringByAddingURLQueryParameters:`
 - Sort URL query keys alphabetically.
+- Removed `ESAppPreviousVersionUserDefaultsKey` defination.
 - Added `ESStringFromNetworkReachabilityStatus(AFNetworkReachabilityStatus status)` function.
 
 ## 3.9.0 (2019-06-05)